update more code examples
