+++ docs/tsconfig.json
@@ -12,7 +12,6 @@
     },
     "include": [
         ".vitepress/**/*.ts",
-        "./examples.ts",
         "./examples/**/*.ts"
     ],
     "exclude": [
